vlmui/models/biomedgpt_model/biomedgpt-requirements.txt
torch==2.5.0
torchvision==0.20.0
pillow==11.1.0

